+++ src/createSinking.ts
@@ -1,42 +1,32 @@
 import { connect, InferableComponentEnhancerWithProps } from 'react-redux';
-import { Sink } from './Sink';
 import { SinkContainer } from './SinkContainer';
-import { Constructor } from './typings';
-import { mergeDispatchState } from './utils';
+import { Constructor, SinkSubscriber } from './typings';
+import { mergeDispatchState, reduceKeys } from './utils';
 
 export function createSinking(container: SinkContainer) {
-  return function<TStateProps = any, TOwnProps = any>(...sinks: Array<Constructor>) {
-    const sinkPrototypes = sinks.map(sink => container.getSinkPrototype(sink));
-    return connect<TStateProps, any, TOwnProps>(
-      createMapStateToProps(sinkPrototypes),
-      createMapDispatchToProps(sinkPrototypes),
-      createMergeProps(sinkPrototypes)
-    ) as InferableComponentEnhancerWithProps<TStateProps, TOwnProps>;
-  };
-}
+  return function <TSink, TStateProps = any, TOwnProps = any>(
+    sink: Constructor<TSink>,
+    subscriber: boolean | SinkSubscriber<TSink> = true
+  ) {
+    const sinkPrototype = container.getSinkPrototype(sink);
+    const namespace = sinkPrototype.namespace;
+    const mapDispatch = () => ({ [namespace]: sinkPrototype.dispatches });
+    let mapState;
+    if (subscriber) {
+      if (typeof subscriber === 'function') {
+        const subscribes = subscriber(sinkPrototype.stateNames as any) as Array<string>;
+        mapState = (state) => ({
+          [sinkPrototype.namespace]: reduceKeys(subscribes, key => state[sinkPrototype.namespace][key])
+        });
+      } else {
+        mapState = (state) => ({ [sinkPrototype.namespace]: state[sinkPrototype.namespace] });
+      }
+    }
+    const mergeProps = (state: any, dispatch: any, own: any) => ({
+      [namespace]: mergeDispatchState(dispatch[namespace], state[namespace]),
+      ...own
+    });
 
-function createMapStateToProps(sinks: Array<Sink>) {
-  return function(state: any) {
-    return sinks.reduce((accumulate: any, sink) => {
-      accumulate[sink.namespace] = state && state[sink.namespace];
-      return accumulate;
-    }, {});
-  };
-}
-
-function createMapDispatchToProps(sinks: Array<Sink>) {
-  return function() {
-    return sinks.reduce((accumulate: any, sink) => (
-      accumulate[sink.namespace] = sink.dispatches, accumulate
-    ), {});
-  };
-}
-
-function createMergeProps(sinks: Array<Sink>) {
-  return function(stateProps: any, dispatchProps: any, ownProps: any) {
-    return sinks.reduce((accumulate, sink) => {
-      accumulate[sink.namespace] = mergeDispatchState(dispatchProps[sink.namespace], stateProps[sink.namespace]);
-      return accumulate;
-    }, { ...ownProps });
+    return connect(mapState, mapDispatch, mergeProps) as InferableComponentEnhancerWithProps<TStateProps, TOwnProps>;
   };
 }