+++ src/createUseSink.ts
@@ -1,14 +1,22 @@
 import { useSelector } from 'react-redux';
 
 import { SinkContainer } from './SinkContainer';
-import { Constructor } from './typings';
-import { mergeDispatchState } from './utils';
+import { Constructor, SinkSubscriber } from './typings';
+import { mergeDispatchState, reduceKeys } from './utils';
 
 export function createUseSink(container: SinkContainer) {
-  return <T>(sink: Constructor<T>, subscribe: boolean = true): T => {
-    if (subscribe) {
+  return <T>(sink: Constructor<T>, subscriber: boolean | SinkSubscriber<T> = true): T => {
+    if (subscriber) {
       const sinkPrototype = container.getSinkPrototype(sink);
-      const sinkState = useSelector<any, T>(state => state[sinkPrototype.namespace]);
+      let sinkState;
+      if (typeof subscriber === 'function') {
+        const subscribes = subscriber(sinkPrototype.stateNames as any) as Array<string>;
+        sinkState = useSelector<any, any>(state =>
+          reduceKeys(subscribes, key => state[sinkPrototype.namespace][key])
+        );
+      } else {
+        sinkState = useSelector<any, T>(state => state[sinkPrototype.namespace]);
+      }
       return mergeDispatchState<T>(sinkPrototype.dispatches, sinkState);
     }
     return container.getSink(sink);